requirements.txt
discord.py==2.4.0
PyYAML==6.0.2
requests==2.32.3